Add version 2.4.3

+++ recipes/libharu/all/conanfile.py
@@ -60,6 +60,8 @@ class LibharuConan(ConanFile):
         tc = CMakeToolchain(self)
         tc.variables["LIBHPDF_SHARED"] = self.options.shared
         tc.variables["LIBHPDF_STATIC"] = not self.options.shared
+        # Honor BUILD_SHARED_LIBS from conan_toolchain (see https://github.com/conan-io/conan/issues/11840)
+        tc.cache_variables["CMAKE_POLICY_DEFAULT_CMP0077"] = "NEW"
         tc.generate()
         tc = CMakeDeps(self)
         tc.generate()
@@ -70,7 +72,7 @@ class LibharuConan(ConanFile):
         cmake.configure()
         cmake.build()
 
-    def _extract_license(self):
+    def _v230_extract_license(self):
         readme = load(save, os.path.join(self.source_folder, "README"))
         match = next(re.finditer("\n[^\n]*license[^\n]*\n", readme, flags=re.I | re.A))
         return readme[match.span()[1]:].strip("*").strip()
@@ -82,15 +84,21 @@ class LibharuConan(ConanFile):
         rm(self, "CHANGES", os.path.join(self.package_folder))
         rm(self, "INSTALL", os.path.join(self.package_folder))
         rm(self, "README", os.path.join(self.package_folder))
-        rmdir(self, os.path.join(self.package_folder, "if"))
-        save(self, os.path.join(self.package_folder, "licenses", "LICENSE"), self._extract_license())
+
+        if Version(self.version) == "2.3.0":
+            rmdir(self, os.path.join(self.package_folder, "if"))
+            save(self, os.path.join(self.package_folder, "licenses", "LICENSE"), self._v230_extract_license())
+        else:
+            copy(self, pattern="LICENSE", dst=os.path.join(self.package_folder, "licenses"), src=self.source_folder)
 
     def package_info(self):
         libprefix = "lib" if is_msvc(self) else ""
-        libsuffix = "{}{}".format(
-            "" if self.options.shared else "s",
-            "d" if is_msvc(self) and self.settings.build_type == "Debug" else "",
-        )
+        libsuffix = ""
+        if Version(self.version) == "2.3.0":
+            libsuffix = "{}{}".format(
+                "" if self.options.shared else "s",
+                "d" if is_msvc(self) and self.settings.build_type == "Debug" else "",
+            )
         self.cpp_info.libs = [f"{libprefix}hpdf{libsuffix}"]
         if self.settings.os == "Windows" and self.options.shared:
             self.cpp_info.defines = ["HPDF_DLL"]
